Ensure host-specific directory exists first

+++ bootstrap.sh
@@ -32,4 +32,4 @@ function install {
 
 dfroot="$(dirname $(realpath $0))"
 install $dfroot/files
-install $dfroot/files_$(hostname)
+[[ -e $dfroot/files_$(hostname) ]] && install $dfroot/files_$(hostname)
